pyproject.toml
[tool.setuptools]
packages = ["blog2epub"]

[tool.poetry]
name = "blog2epub"
version = "1.5.0"
description = "Convert blog (blogspot.com, wordpress.com or another based on Wordpress) to epub using CLI or GUI."
authors = ["Bohdan Bobrowski <bohdanbobrowski@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pydantic = "^2.8.2"
beautifulsoup4 = "^4.12.3"
ebooklib = "^0.18"
python-dateutil = "^2.9.0.post0"
atoma = "^0.0.17"
fake-useragent = "^1.5.1"
pyyaml = "^6.0.1"
kivy = "^2.3.0"
kivymd = "^1.2.0"
pydantic-yaml = "^1.3.0"
plyer = "^2.1.0"
pyjnius = "^1.6.1"
strip-tags = "^0.5.1"
imagesize = "^1.4.1"
filetype = "^1.2.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pyinstaller = "^6.8.0"
mypy = "^1.10.1"
pytest = "^8.3.2"
lxml-stubs = "^0.5.1"
types-pyyaml = "^6.0.12.20240808"
types-pyinstaller = "^6.10.0.20240812"
types-python-dateutil = "^2.9.0.20240821"
types-requests = "^2.32.0.20240712"
cython = "^3.0.11"
pytest-cov = "^6.0.0"

[tool.poetry.scripts]
blog2epub = "blog2epub.blog2epub_cli:main"
blog2epubgui = "blog2epub.blog2epub_gui:main"
build_gui_windows = "blog2epub_build:build_gui_windows"
build_gui_macos = "blog2epub_build:build_gui_macos"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "I"]
ignore = ["E501"]

[tool.mypy]
exclude = ["venv", "dist", "build"]


[tool.ruff.lint.extend-per-file-ignores]
"blog2epub_gui.py" = ["E402"]

# TODO: verify if this default value bolow can be changed to setupotools
# [build-system]
# requires = ["poetry-core"]
# build-backend = "poetry.core.masonry.api"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"